cleanup: Remove the syntax for multi-declarators.
We no longer allow `int a, b;` or any variant of that declaration. No
support for multiple declarators per declaration. It turns out that most
instances of this syntax was a bad idea in the first place (so it's good
to disallow), and the few that were kind of useful were maybe 3-4 and it
didn't add much value to support this syntax given its potential (and
real! in toxcore!) abuses.