cleanup: Wrap all compound statements in CompoundStmt.
This makes it easier to identify and distinguish from other Node lists
such as top level decls and struct/union/enum members. I also changed
SwitchStmt to only accept `case` and `default`. Its body is not
considered a CompoundStmt, but rather a list of switch cases. We can do
this now because we enforce that all cases have a single (usually
compound) statement as their body.