refactor: Slightly better AST representation of decls.
This pulls all the type-related information into a single AST node,
making typechecking easier later on. Also, restricted `#ifdef/#ifndef` to
only allow `#else`, no `#elif`. If you want to use `#elif`, use
`#if defined` instead of `#ifdef`. This may make typechecking easier as
well because structs with ifdefs in them can rely on a binary either-or
instead of a complicated conditional for field presence.